% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuRnet.R
\name{cuRnet_bfs}
\alias{cuRnet_bfs}
\title{Breadth-first search. 
This function traverses the graph via a breadth-first search from a given set of source vertices, and returns depth of visited nodes.}
\usage{
cuRnet_bfs(graph, sources)
}
\arguments{
\item{graph}{A cuRnet graph object created with \code{cuRnet_graph}.}

\item{sources}{The lists of source vertices from which to start BFSs. Per every source, one BFS is performed.}
}
\value{
A \code{NumericMatrix} having a number of rows equal to the number of source vertices, and a number of columns equal to the total number of vertices of th einput graph. Every row correspondo to a specific source vertex, and row cell reports the depth from the given source to the correspondig graph vertex.
}
\description{
Breadth-first search. 
This function traverses the graph via a breadth-first search from a given set of source vertices, and returns depth of visited nodes.
}
\details{

}
\examples{
\dontrun{
library(igraph)
library(cuRnet)
rg <- sample_fitness_pl(100, 1000, 2.2, 2.3)
cdf <- data.frame( ends(rg, E(rg))[,1], ends(rg, E(rg))[,2] )
colnames(cdf) <- c("from", "to")
sources <- union(cdf$from, cdf$to)[1:20]
cg <- cuRnet_graph(cdf)
bfs <- cuRnet_bfs(cg, sources)
bfs[1,]
}

}
