% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuRnet.R
\name{cuRnet_sssp}
\alias{cuRnet_sssp}
\title{Single Source Shortest Paths: distances and predecessors.
This function computes shortest paths from a series of vertices.
For each source vertex, shortest paths to every vertex in the graph are computed.
The function returns distances and predecessors.}
\usage{
cuRnet_sssp(graph, from)
}
\arguments{
\item{graph}{A cuRnet graph object created with \code{cuRnet_graph}.}

\item{from}{A \code{CharacterVector} with the names of the source vertices.}
}
\value{
A list of two \code{NumericMatrix} indexed by "distances" and "predecessors". Rows are source vertices and colums are network vertices. A entry is: "distances" the distance along the shortest path from the source to the destination vertex; "predecessors" a minimal predecessor of the vertex and along the shortest path.
}
\description{
Single Source Shortest Paths: distances and predecessors.
This function computes shortest paths from a series of vertices.
For each source vertex, shortest paths to every vertex in the graph are computed.
The function returns distances and predecessors.
}
\details{

}
\examples{
\dontrun{
library(STRINGdb)
library(igraph)
library(cuRnet)
ss <- STRINGdb$new( version="10", species=9606, score_threshold=900)
g <- ss$get_graph()
from <- V(g)$name[1:10]
x <- data.frame("from" = ends(g,E(g))[,1], "to" = ends(g,E(g))[,2], "score" = E(g)$combined_score/1000)
cg <- cuRnet_graph(x)
ret <- cuRnet_sssp(g, from)
ret[["distances"]]
ret[["predecessors"]]
}

}
