% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuRnet.R
\name{cuRnet_sssp_dists}
\alias{cuRnet_sssp_dists}
\title{Single Source Shortest Paths: distances only.
This function computes shortest paths from a series of vertices.
For each source vertex, shortest paths to every vertex in the graph are computed.
The function returns only distances.}
\usage{
cuRnet_sssp_dists(graph, from)
}
\arguments{
\item{graph}{A cuRnet graph object created with \code{cuRnet_graph}.}

\item{from}{A \code{CharacterVector} with the names of the source vertices.}
}
\value{
A \code{NumericMatrix} where rows are source vertices and colums are network vertices. An entry is the distance along the shortest path from the source to the destination vertex.
}
\description{
Single Source Shortest Paths: distances only.
This function computes shortest paths from a series of vertices.
For each source vertex, shortest paths to every vertex in the graph are computed.
The function returns only distances.
}
\details{

}
\examples{
\dontrun{
library(STRINGdb)
library(igraph)
library(cuRnet)
ss <- STRINGdb$new( version="10", species=9606, score_threshold=900)
g <- ss$get_graph()
from <- V(g)$name[1:10]
x <- data.frame("from" = ends(g,E(g))[,1], "to" = ends(g,E(g))[,2], "score" = E(g)$combined_score/1000)
cg <- cuRnet_graph(x)
ret <- cuRnet_sssp_dists(cg, from)
ret[["distances"]][1,] 
ret[["predecessors"]][1,]
ret[1,]
}

}
